--- LocalFootball/TeamStatistic+CoreDataClass.swift
@@ -1,48 +1,0 @@
-//
-//  TeamStatistic+CoreDataClass.swift
-//  LocalFootball
-//
-//  Created by Дарья Леонова on 12.04.2020.
-//  Copyright © 2020 Дарья Леонова. All rights reserved.
-//
-//
-
-import Foundation
-import CoreData
-import SwiftyJSON
-
-@objc(TeamStatistic)
-public class TeamStatistic: NSManagedObject, Decodable {
-    
-    enum CodingKeys: String, CodingKey {
-        case fullStatistics = "fullStatistics"
-        case tournamentsStatistics = "tournamentsStatistics"
-    }
-    
-    required convenience public init(from decoder: Decoder) throws {
-        guard let contexUserInfoKey = CodingUserInfoKey.context,
-            let managedObjectContext = decoder.userInfo[contexUserInfoKey] as? NSManagedObjectContext,
-            let entity = NSEntityDescription.entity(forEntityName: "TeamStatistic", in: managedObjectContext) else { fatalError("Failed to decode TeamStatistic") }
-        self.init(entity: entity, insertInto: managedObjectContext)
-        
-        let values = try decoder.container(keyedBy: CodingKeys.self)
-        do {
-            fullStatistics = try values.decode(Statistics.self, forKey: .fullStatistics)
-            let tournamentStatistics = try values.decode([TournamentStatistics].self, forKey: .tournamentsStatistics)
-            tournamentStatistics.forEach { self.addToTournamentStatistics($0) }
-            
-        } catch let error as NSError {
-            print(error.localizedDescription)
-        }
-    }
-    
-    func update(with teamStatisticsJSON: JSON) {
-        self.fullStatistics?.update(with: teamStatisticsJSON["fullStatistics"])
-        var i = 0
-        for tournamentStatistics in self.tournamentsStatistics! {
-            (tournamentStatistics as? TournamentStatistics)?.update(with: teamStatisticsJSON["tournamentStatistics"].arrayValue[i])
-            i += 1
-        }
-    }
-    
-}
